+++ src/refactorings/lift-up-conditional/lift-up-conditional.ts
@@ -2,9 +2,9 @@ import { Editor, ErrorReason } from "../../editor/editor";
 import { Selection } from "../../editor/selection";
 import * as t from "../../ast";
 
-export { bubbleUpIfStatement, createVisitor as canBubbleUpIfStatement };
+export { liftUpConditional, createVisitor as canliftUpConditional };
 
-async function bubbleUpIfStatement(editor: Editor) {
+async function liftUpConditional(editor: Editor) {
   const { code, selection } = editor;
   const updatedCode = updateCode(t.parse(code), selection);